refactor(python-sdk): split tasks into separate modules
* Remove duplicated logic.
* Move example code to examples.
